Add Preset superclass to IPCameraPreset
