+++ backend/src/main/java/com/benine/backend/preset/Preset.java
@@ -3,6 +3,7 @@ package com.benine.backend.preset;
 import com.benine.backend.camera.Camera;
 import com.benine.backend.camera.CameraBusyException;
 import com.benine.backend.camera.CameraConnectionException;
+import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
 import java.util.Collection;
@@ -94,7 +95,20 @@ public abstract class Preset {
    *
    * @return JSON representation of this object.
    */
-  public abstract JSONObject toJSON();
+  public JSONObject toJSON() {
+    JSONObject json = new JSONObject();
+
+    json.put("id", getId());
+    json.put("cameraid", getCameraId());
+    json.put("image", getImage());
+    JSONArray tagsJSON = new JSONArray();
+    for (String tag : tags) {
+      tagsJSON.add(tag);
+    }
+    json.put("tags", tagsJSON);
+
+    return json;
+  }
 
 
   @Override
@@ -152,4 +166,5 @@ public abstract class Preset {
    * @return the query.
    */
   public abstract String createDeleteSQL();
+
 }
